+++ packages/review/src/components/infinity-list/index.ts
@@ -1,0 +1,1 @@
+export * from './popular-reviews-infinite'
